requirements.txt
flask
pandas
numpy
tvDatafeed
plotly
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
scikit-learn
